Ylw-3314 refactor(CategoryController): chand to SingleOrDefault
